+++ src/stream.ts
@@ -1,4 +1,3 @@
-import { abortableSource } from 'abortable-iterator'
 import { encode, decode } from 'it-length-prefixed'
 import { pipe } from 'it-pipe'
 import { pushable, type Pushable } from 'it-pushable'
@@ -16,26 +15,26 @@ interface InboundStreamOpts {
 }
 
 export class OutboundStream {
-  private readonly pushable: Pushable<Uint8Array>
-  private readonly lpPushable: Pushable<Uint8ArrayList>
+  private readonly pushable: Pushable<Uint8Array | Uint8ArrayList>
   private readonly closeController: AbortController
   private readonly maxBufferSize: number
 
   constructor (private readonly rawStream: Stream, errCallback: (e: Error) => void, opts: OutboundStreamOpts) {
-    this.pushable = pushable({ objectMode: false })
-    this.lpPushable = pushable({ objectMode: false })
+    this.pushable = pushable()
     this.closeController = new AbortController()
     this.maxBufferSize = opts.maxBufferSize ?? Infinity
 
+    this.closeController.signal.addEventListener('abort', () => {
+      rawStream.close()
+        .catch(err => {
+          rawStream.abort(err)
+        })
+    })
+
     pipe(
-      abortableSource(this.pushable, this.closeController.signal, { returnOnAbort: true }),
-      (source) => encode(source),
+      this.pushable,
       this.rawStream
     ).catch(errCallback)
-
-    pipe(abortableSource(this.lpPushable, this.closeController.signal, { returnOnAbort: true }), this.rawStream).catch(
-      errCallback
-    )
   }
 
   get protocol (): string {
@@ -49,25 +48,23 @@ export class OutboundStream {
       throw Error(`OutboundStream buffer full, size > ${this.maxBufferSize}`)
     }
 
-    this.pushable.push(data)
+    this.pushable.push(encode.single(data))
   }
 
   /**
    * Same to push() but this is prefixed data so no need to encode length prefixed again
    */
   pushPrefixed (data: Uint8ArrayList): void {
-    if (this.lpPushable.readableLength > this.maxBufferSize) {
+    if (this.pushable.readableLength > this.maxBufferSize) {
       throw Error(`OutboundStream buffer full, size > ${this.maxBufferSize}`)
     }
-    this.lpPushable.push(data)
+    this.pushable.push(data)
   }
 
   async close (): Promise<void> {
     this.closeController.abort()
     // similar to pushable.end() but clear the internal buffer
     await this.pushable.return()
-    await this.lpPushable.return()
-    await this.rawStream.close()
   }
 }
 
@@ -81,17 +78,20 @@ export class InboundStream {
     this.rawStream = rawStream
     this.closeController = new AbortController()
 
-    this.source = abortableSource(
-      pipe(this.rawStream, (source) => decode(source, opts)),
-      this.closeController.signal,
-      {
-        returnOnAbort: true
-      }
+    this.closeController.signal.addEventListener('abort', () => {
+      rawStream.close()
+        .catch(err => {
+          rawStream.abort(err)
+        })
+    })
+
+    this.source = pipe(
+      this.rawStream,
+      (source) => decode(source, opts)
     )
   }
 
   async close (): Promise<void> {
     this.closeController.abort()
-    await this.rawStream.close()
   }
 }